Correct BQ24210DQCT BAT pin to Power Input.
Realised it was incorrectly specified as a Power Output while porting Notecarrier-A. It’s the battery itself which is the source of power, so even though this pin provides charging current, it must be a Power Input since it is connected to the Power Output of the battery. The symbol has been fixed, so I updated it and put the PWR_FLAG where it should be - at the battery.