Backport name change from `RS-0402` to `RS-0402_NO-FILL`.
Name changed in Notecarrier-A port. See those PortingNotes for details. Also performed long overdue “update PCB from schematic” - remember that the test points are not in the schematic, so must be preserved in the PCB! The update has no material effect, only updates to library references. Confirmed by comparing Gerbers before and after - only change is the fiducial mask layer has slightly smaller radius. The difference is not significant.