Set all testpoints to “Not in schematic”.
After backporting improvements from the Notecarrier-A port, I realised that updating the PCB from the schematic can delete all the testpoints, because they have been added at the PCB level. In Notecarrier-A I opted to add them to the schematic, since there was only 2, so it was easy to bring it inline with common KiCad practice. In Notecarrier-B, where there are 34, rather than adding them all to the schematic, I opted to simply mark them as “Not in schematic”. Now KiCad knows not to expect them to be in the schematic, and will not delete them if the schematic is synced with the PCB.