Complete validation.
A few minor touch-ups to fills, polygons and DRC exclusions.
Most significantly, re-assigned “GND” net to most (all?) of the weird-no-fill-via footprints, which got unassigned in a previous update-from-libraries. Also locked them all to reduce likely of losing the assignments in future, but it’s a hack so not sure it will stick - be wary. Also locked the M2 footprint because I overrode the fill settings on a few pads to “none” to prevent thermal reliefs being generated.